# Conflicts:
#	lib/screens/account_screen.dart
#	lib/services/account_service.dart

+++ lib/services/api_client.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:flutter/cupertino.dart';
 import 'package:frontend/services/keycloak_service.dart';
 import 'package:frontend/state/api_selection_provider.dart';
 import 'package:http/http.dart' as http;
@@ -14,8 +15,9 @@ class ApiClient {
     return ApiClient._create(backendUrl);
   }
 
-  Future<http.Response> authorizedGet(String endpoint) async {
-    final accesstoken = await KeycloakService().getAccessToken();
+  Future<http.Response?> authorizedGet(BuildContext context, String endpoint) async {
+    final accesstoken = await KeycloakService().getAccessToken(context);
+    if (accesstoken == null) return null;
     final response = await http.get(
       Uri.parse('$_backendUrl/$endpoint'),
       headers: {
@@ -38,8 +40,9 @@ class ApiClient {
     return response;
   }
 
-  Future<http.Response> authorizedPost(String endpoint, Map<String, dynamic> body) async {
-    final accesstoken = await KeycloakService().getAccessToken();
+  Future<http.Response?> authorizedPost(BuildContext context, String endpoint, Map<String, dynamic> body) async {
+    final accesstoken = await KeycloakService().getAccessToken(context);
+    if (accesstoken == null) return null;
     final response = await http.post(
       Uri.parse('$_backendUrl/$endpoint'),
       headers: {
@@ -66,8 +69,9 @@ class ApiClient {
     return response;
   }
 
-  Future<http.Response> authorizedPut(String endpoint, Map<String, dynamic> body) async {
-    final accesstoken = await KeycloakService().getAccessToken();
+  Future<http.Response?> authorizedPut(BuildContext context, String endpoint, Map<String, dynamic> body) async {
+    final accesstoken = await KeycloakService().getAccessToken(context);
+    if (accesstoken == null) return null;
     final response = await http.put(
       Uri.parse('$_backendUrl/$endpoint'),
       headers: {
@@ -94,8 +98,9 @@ class ApiClient {
     return response;
   }
 
-  Future<http.Response> authorizedDelete(String endpoint) async {
-    final accesstoken = await KeycloakService().getAccessToken();
+  Future<http.Response?> authorizedDelete(BuildContext context, String endpoint) async {
+    final accesstoken = await KeycloakService().getAccessToken(context);
+    if (accesstoken == null) return null;
     final response = await http.delete(
       Uri.parse('$_backendUrl/$endpoint'),
       headers: {
@@ -108,8 +113,9 @@ class ApiClient {
     return response;
   }
 
-  Future<http.Response> authorizedDeleteWithBody(String endpoint, Map<String, dynamic> body) async {
-    final accesstoken = await KeycloakService().getAccessToken();
+  Future<http.Response?> authorizedDeleteWithBody(BuildContext context, String endpoint, Map<String, dynamic> body) async {
+    final accesstoken = await KeycloakService().getAccessToken(context);
+    if (accesstoken == null) return null;
     final response = await http.delete(
       Uri.parse('$_backendUrl/$endpoint'),
       headers: {
